# docs/openapi/swagger.yaml
openapi: 3.1.0
info:
  title: FGD (Federation Governance Dashboard) API
  version: "2.1.0"
  description: |
    REST interface for FGD, the Minecraft NPC swarm management system.
    Routes are mounted as follows:
      - Unversioned surface: `/api/*` — currently served by the v1 route
        tree. `/api/mineflayer` is a special case that always resolves to
        the policy-gated v2 Mineflayer surface (or 503 if the policy
        service failed to initialize).
      - `/api/v1/*` — legacy direct-control surface. Mineflayer endpoints
        under `/api/v1/mineflayer` are DEPRECATED (sunset 2026-07-01).
      - `/api/v2/*` — current surface, including the policy-gated
        Mineflayer routes at `/api/v2/mineflayer`.

    Authentication: protected routes accept either a Bearer JWT
    (`Authorization: Bearer <token>`) or an API key (`X-API-Key`).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Position:
      type: object
      properties:
        x: { type: number }
        y: { type: number }
        z: { type: number }
      required: [x, y, z]
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/npcs:
    get:
      tags: [NPCs]
      summary: List NPCs with pagination
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: Paged NPC list }
    post:
      tags: [NPCs]
      summary: Create an NPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                role: { type: string }
                npcType: { type: string }
                appearance: { type: object }
                personality: { type: object }
                position: { $ref: '#/components/schemas/Position' }
                autoSpawn: { type: boolean, default: false }
      responses:
        '201': { description: Created }
        '400': { description: Missing role or npcType }
  /api/npcs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [NPCs]
      summary: Fetch one NPC (optionally enriched with learning profile)
      responses:
        '200': { description: NPC found }
        '404': { description: Not found }
    put:
      tags: [NPCs]
      summary: Update NPC profile
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      tags: [NPCs]
      summary: Finalize an NPC (archive and optionally remove from world)
      parameters:
        - in: query
          name: preserve
          schema: { type: boolean }
        - in: query
          name: removeFromWorld
          schema: { type: boolean }
      responses:
        '200': { description: Finalized }
...